ethdb: Move testsuite to dbtest, fix lints
